+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## main
 
+* Experimental: add `ViewComponent::Styleable` for encapsulated CSS.
+
+    *Joel Hawksley*
+
 * Experimental: call `._sidecar_files` to fetch the sidecar files for a given list of extensions, e.g. passing `["yml", "yaml"]`.
 
     *Elia Schito*